// src/client/domain/InventoryView.ts
export class InventoryView {

    id?: string;
    productCode: string;
    productName: string;
    operationType?: string;
    invoiceDetailId?: string;
    description?: string;
    quantity: number;
    unitPrice: number;
    discountAmount?: number;
    operationDate: string;
    total?: number;
}